+++ android/filament-android/src/main/java/com/google/android/filament/IndirectLight.java
@@ -288,7 +288,7 @@ public class IndirectLight {
          * range can be adjusted with this method.</p>
          *
          * @param envIntensity  Scale factor applied to the environment and irradiance such that
-         *                      the result is in cd/m^2 (lux) units (default = 30000)
+         *                      the result is in <i>lux</i>, or <i>lumen/m^2</i> (default = 30000)
          *
          * @return This Builder, for chaining calls.
          */
@@ -354,14 +354,14 @@ public class IndirectLight {
      * range can be adjusted with this method.</p>
      *
      * @param intensity  Scale factor applied to the environment and irradiance such that
-     *                   the result is in cd/m^2 units (default = 30000)
+     *                   the result is in <i>lux</i>, or <i>lumen/m^2</i> (default = 30000)
      */
     public void setIntensity(float intensity) {
         nSetIntensity(getNativeObject(), intensity);
     }
 
     /**
-     * Returns the environment's intensity in cd/m<sup>2</sup>.
+     * Returns the environment's intensity in <i>lux</i>, or <i>lumen/m^2</i>.
      */
     public float getIntensity() {
         return nGetIntensity(getNativeObject());