+++ filament/src/details/Renderer.h
@@ -31,6 +31,8 @@
 #include <filament/Renderer.h>
 #include <filament/View.h>
 
+#include <fg2/FrameGraphId.h>
+
 #include <backend/DriverEnums.h>
 #include <backend/Handle.h>
 #include <backend/PresentCallable.h>
@@ -124,7 +126,9 @@ private:
     friend class Renderer;
     using Command = RenderPass::Command;
 
-    backend::Handle<backend::HwRenderTarget> getRenderTarget(FView& view) const noexcept;
+    void getRenderTarget(FView const& view,
+            backend::TargetBufferFlags& outAttachementMask,
+            backend::Handle<backend::HwRenderTarget>& outTarget) const noexcept;
 
     void readPixels(backend::Handle<backend::HwRenderTarget> renderTargetHandle,
             uint32_t xoffset, uint32_t yoffset, uint32_t width, uint32_t height,