+++ libs/gltfio/src/ResourceLoader.cpp
@@ -256,11 +256,22 @@ void ResourceLoader::addResourceData(const char* uri, BufferDescriptor&& buffer)
         SYSTRACE_CONTEXT();
         SYSTRACE_ASYNC_BEGIN("addResourceData", 1);
     }
+    // NOTE: replacing an existing item in a robin map does not seem to behave as expected.
+    // To work around this, we explicitly erase the old element if it already exists.
+    auto iter = pImpl->mUriDataCache.find(uri);
+    if (iter != pImpl->mUriDataCache.end()) {
+        pImpl->mUriDataCache.erase(iter);
+    }
     pImpl->mUriDataCache.emplace(uri, std::move(buffer));
 }
 
 bool ResourceLoader::hasResourceData(const char* uri) const {
     return pImpl->mUriDataCache.find(uri) != pImpl->mUriDataCache.end();
+}
+
+void ResourceLoader::evictResourceData() {
+    // Note that this triggers BufferDescriptor callbacks.
+    pImpl->mUriDataCache.clear();
 }
 
 bool ResourceLoader::loadResources(FilamentAsset* asset) {